fix(hype-ui) Small Ui fixes
