+++ services/client/src/containers/redeem/index.ts
@@ -1,0 +1,2 @@
+export * from './RewardsListContainer';
+export * from './TransactionsContainer';
