+++ services/client/src/models/Reward.model.ts
@@ -1,0 +1,10 @@
+import { TransactionStatus } from "../utils";
+
+export type TransactionItem = {
+  value: number;
+  pool?: string;
+  status: TransactionStatus;
+  startDate: Date;
+};
+
+export type Reward = Omit<TransactionItem, 'status'>;
